# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: audit — API fiscal (ESAA-Flow)
  version: 1.0.0
  description: |
    Contrato mínimo para destravar o frontend (Lovable). Toda escrita passa pelo
    orquestrador single-writer: o cliente envia uma *intenção*; a API valida
    (7 camadas), grava o evento e re-projeta. Leituras vêm da projeção.

    Convenções:
    - Identificadores em inglês; mensagens de erro em PT-BR.
    - `cnpj` sempre com 14 dígitos, sem máscara.
    - `period` no formato `YYYY-MM` (competência).
    - Toda resposta de escrita devolve o `event_seq` gravado e o `projection_hash`.
servers:
  - url: http://localhost:3000/v1
    description: desenvolvimento
  - url: https://api.SEU-DOMINIO.com.br/v1
    description: produção
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [auth]
      security: []
      summary: Login do usuário do escritório
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        "200":
          description: Token JWT
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  expires_in: { type: integer, example: 3600 }
                  tenant: { $ref: "#/components/schemas/Tenant" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Token ausente ou inválido
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Tenant:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: "Escritório Exemplo Contabilidade" }
        plan: { type: string }
...
    Error:
      type: object
      properties:
        code: { type: string }
        message: { type: string }